Trait tokkit::metrics::MetricsCollector

pub trait MetricsCollector {
    fn incoming_introspection_request(&self);

    fn introspection_request(&self, request_started: Instant);

    fn introspection_request_success(&self, request_started: Instant);

    fn introspection_request_failure(&self, request_started: Instant);

    fn introspection_service_call(&self, request_started: Instant);

    fn introspection_service_call_failure(&self, request_started: Instant);

    fn introspection_service_call_success(&self, request_started: Instant);
}

Collects metrics for token introspection

Required methods

fn incoming_introspection_request(&self)

An incoming request for token introspection

fn introspection_request(&self, request_started: Instant)

The complete introspection workflow was finished regardless of the result.

fn introspection_request_success(&self, request_started: Instant)

The complete introspection workflow was finished and successful

fn introspection_request_failure(&self, request_started: Instant)

The complete introspection workflow was finished and failed

fn introspection_service_call(&self, request_started: Instant)

The token introspections was called regardless of the result.

fn introspection_service_call_failure(&self, request_started: Instant)

The token introspections was called and the call was a failure.

fn introspection_service_call_success(&self, request_started: Instant)

The token introspections was called and the call was a success.

Loading content...