Struct tokkit::parsers::CustomTokenInfoParser

pub struct CustomTokenInfoParser {
    pub active_field: Option<String>,
    pub user_id_field: Option<String>,
    pub scope_field: Option<String>,
    pub expires_in_field: Option<String>,
}

A configurable TokenInfoParser that parses a TokenInfo from JSON returned by a token introspection service.

Fields
active_field: Option<String>

The field name in the JSON that identifies the active field for the TokenInfo. If None the field will not be looked up and set to true in the TokenInfo right away. The reason is that this is a mandatory field in the TokenInfo and that we assume that if a token introspection service does not provide this field it would return an error in the introspection request in case the token is not active at the time the request was made.

user_id_field: Option<String>

The field name in the JSON that identifies the user_id field for the TokenInfo. If None the field will not be looked up and set to None in the TokenInfo right away.

scope_field: Option<String>

The field name in the JSON that identifies the scope field for the TokenInfo. If None the field will not be looked up and set to None in the TokenInfo right away.

expires_in_field: Option<String>

The field name in the JSON that identifies the expires_in field for the TokenInfo. If None the field will not be looked up and set to None in the TokenInfo right away.

Implementations

impl CustomTokenInfoParser

pub fn new<U, S, E, A>(
active_field: Option<A>,
user_id_field: Option<U>,
scope_field: Option<S>,
expires_in_field: Option<E>
) -> Self where
U: Into<String>,
S: Into<String>,
E: Into<String>,
A: Into<String>,

pub fn from_env() -> Result<CustomTokenInfoParser, Error>

Create a new parser from environment variables.

The following variables used to identify the field in a token info response:

TOKKIT_TOKEN_INFO_PARSER_USER_ID_FIELD(optional): The field name for the user id * TOKKIT_TOKEN_INFO_PARSER_SCOPE_FIELD(optional): The field name for scopes

TOKKIT_TOKEN_INFO_PARSER_EXPIRES_IN_FIELD(optional): The field name for the * TOKKIT_TOKEN_INFO_PARSER_ACTIVE_FIELD(optional): The field name for the active field