Module tokkit::token_manager

Managing AccessTokens.

AccessTokens are managed by configuring a ManagedToken. They can later be queried by the identifier configured with the ManagedToken. The identifier can be any type T where T: Eq + Ord + Send + Sync + Clone + Display + 'static

Modules

token_provider
Interaction with the authorization server

Structs

AccessTokenManager
The TokenManager refreshes AccessTokenss in the background.

AccessTokenSource

AccessTokenSourceSync
An AccessTokenSource with the Sync trait.

FixedAccessTokenSource

FixedAccessTokenSourceSync
A source for fixed access tokens which implements the Sync trait

ManagedToken
An AccessToken to be managed. The AccessToken will be updated automatically.

ManagedTokenBuilder
A builder to configure a ManagedToken.

ManagedTokenGroup
A group of ManagedTokens that are requested from the same authorization server

ManagedTokenGroupBuilder

TokenError

Enums

TokenErrorKind

Traits

GivesAccessTokensById
Can be queired for AccessTokens by their identifier configured with the respective ManagedToken.

GivesFixedAccessToken
Can be queried for a fixed AccessToken.

Type Definitions

TokenResult