Struct tokkit::token_manager::ManagedTokenGroupBuilder

pub struct ManagedTokenGroupBuilder<T, S: AccessTokenProvider + 'static> { /* fields omitted */ }

Implementations

impl<T: Eq + Send + Clone + Display, S: AccessTokenProvider + Send + Sync + 'static> ManagedTokenGroupBuilder<T, S>

pub fn with_token_provider(&mut self, token_provider: S) -> &mut Self

Sets the AccessTokenProvider for this group of ManagedTokens. This is a mandatory value.

pub fn with_managed_token(
&mut self,
managed_token: ManagedToken<T>
) -> &mut Self

Adds a ManagedToken to this group.

pub fn with_refresh_threshold(&mut self, refresh_threshold: f32) -> &mut Self

Sets the refresh interval as a percentage of the "expires in" sent by the authorization server. The default is 0.75

pub fn with_warning_threshold(&mut self, warning_threshold: f32) -> &mut Self

Sets the warnoing interval as a percentage of the "expires in" sent by the authorization server. The default is 0.85

pub fn with_managed_token_from_builder(
&mut self,
builder: ManagedTokenBuilder<T>
) -> StdResult<&mut Self, InitializationError>

Adds a ManagedToken built from the given ManagedTokenBuilder.

pub fn single_token(token_id: T, scopes: Vec<Scope>, token_provider: S) -> Self

Sets everything needed to manage the give token.

pub fn single_token_from_env(
token_id: T,
token_provider: S
) -> StdResult<Self, InitializationError>

Sets everything needed to manage the give token.

Ssopes are read from TOKKIT_MANAGED_TOKEN_SCOPES

pub fn build(self) -> StdResult<ManagedTokenGroup<T>, InitializationError>

Build the ManagedTokenGroup.

Fails if not all required fields are set properly.