Module tokkit::token_manager::token_provider

Interaction with the authorization server

Modules

credentials

Structs

AuthorizationRequestError
An error in detail returned by the authorization server.

AuthorizationServerResponse
The response an AccessTokenProvider received from an authorization server.

EnvAccessTokenProvider
Provides access tokens from an environment variable

ResourceOwnerPasswordCredentialsGrantProvider
Provides tokens via Resource Owner Password Credentials Grant

Enums

AccessTokenProviderError

AuthorizationServerErrorCode
The error code returned from the authorization server on BadRequest.

Traits

AccessTokenProvider
Calls an authorization server for an AccessToken and the time left until the AccessToken expires.

Type Definitions

AccessTokenProviderResult