Module tokkit::token_manager::token_provider::credentials

Modules

parsers
Parser for credentials, mostl likely JSON

Structs

ClientCredentials
Credentials of the registered client to POST an Authorization Request

RequestTokenCredentials

ResourceOwnerCredentials
Credentials of the resource Owner required for the Access Token Request

SplitFileCredentialsProvider
Reads the credentials for the resource owner and the client from two seperate (mostly) JSON files.

Enums

CredentialsError

Traits

CredentialsProvider

Type Definitions

CredentialsResult
Type alias for the common return type regarding credentials