Trait tokkit::token_manager::token_provider::AccessTokenProvider

pub trait AccessTokenProvider {
    fn request_access_token(
        &self, 
        scopes: &[Scope]
    ) -> AccessTokenProviderResult;
}

Calls an authorization server for an AccessToken and the time left until the AccessToken expires.

Implementors may use different flows to interact with the authorization server.

Required methods

fn request_access_token(&self, scopes: &[Scope]) -> AccessTokenProviderResult

Issue a request to the authorization server for an AccessToken with the given Scopes.

Loading content...